COA-7 fix: an exception is thrown when a user logs out of a second window of the same browser after previously logging out of the first one already.
- A user could potentially have multiple browser windows open to the site. If the
user logged out in one window, thereby setting 'current_user' to 'nil', clicking
the "Log out" link in a second window would result in an error because of 'forget(current_user)' in
the 'log_out' method. We can avoid this by logging out only if the user is logged in.